+++ changelog.md
@@ -1,4 +1,12 @@
 # Changelog
+#v1.4.2 Changes
+- Backend/Frontend/API: Added festival support. You can add lottie and schedule it.
+- API: Added ServiceLater Support. Service url will be passed to view too.
+- Core: Added module data modifier
+- See how it works:
+- Create class/method `App\Parser\ModuleDataModifier` -> `moduleAliasNameMethod($data, $moduleInfo)`
+  and manipulate your data and return it. Works for API and views.
+- Do not forget to run `php artisan migrate` if you are upgrading.
 
 #1.4.1
 - Backend: Export the correct folder (neo) while installing the package